Create simple webcam capture
If you dont already have it, here is a simple webcam capture if you guys are needing this to do anything with your portion of the project. Sorry I didnt get it up faster. 

I will update again with the functionalities we discussed earlier as soon as I get it done.