Create new_follow_object
This is a different algorithm that I think might work pretty well if we can determine how to find the "center" of the object. This follow algorithm should mimic the movements of that point for however many "steps" we specify (right now, it's at 10). Unfortunately, it does have a couple of for loops in it, but it has no recursion (:D)

I modified the "track()" method to include a "diameter" for the shape we're following. If the shape is a circle, int z is a diameter; if something else, int z is the diameter of a circle that will fit in the shape so that all the shape's sides are touching the edge of the circle, but the shape must be equilateral. Theoretically, this track() should only return true if it is testing the center of the shape.